refactor(gm-bridge): extract GMBridge effects to new gm_bridge_effects.js and update imports
